finetune/requirements.txt
torch>=2.0.0
datasets>=2.13.0
transformers>=4.36.0
unsloth>=2023.12.0
trl>=0.7.1
matplotlib>=3.7.0
pandas>=2.0.0
tqdm>=4.66.0
pillow>=10.0.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
peft>=0.5.0
ray>=2.6.0
requests>=2.28.0
spacy>=3.0.0
google-cloud-storage>=2.0.0
huggingface_hub>=0.16.0 